requirements.txt
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
comtypes==1.4.6
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-xlib==0.33
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
six==1.16.0
tzdata==2024.1
